requirements.txt
langchain
openai
streamlit
python-dotenv
tiktoken
faiss-cpu
PyMuPDF
